+++ src/main/java/lotto/LottoApplication.java
@@ -1,5 +1,6 @@
 package lotto;
 
+import lotto.controller.LottoGame;
 import lotto.domain.*;
 import lotto.view.InputView;
 import lotto.view.OutputView;
